chore: remove publicly exported paypal js types
